backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
PyPDF2==3.0.1
python-docx==0.8.11
aiofiles==23.2.0
python-dotenv==1.0.0
